Rename: component 이름 변경
- SearchVar -> SearchBar (오타로 판단해 변경)
- BarChart -> ChartBar (다른 Chart 컴포넌트와 통일성을 위해)
- Result -> MyResult
